+++ hutool-core/src/main/java/cn/hutool/core/date/DateUtil.java
@@ -867,6 +867,22 @@ public class DateUtil extends CalendarUtil {
 				// 格式类似：2018-09-13T05:34:31+08:00
 				return parse(utcString, DatePattern.UTC_WITH_XXX_OFFSET_FORMAT);
 			}
+		} else if(ReUtil.contains("-\\d{2}:?00", utcString)){
+			// Issue#2612，类似 2022-09-14T23:59:00-08:00 或者 2022-09-14T23:59:00-0800
+
+			// 去除类似2019-06-01T19:45:43 -08:00加号前的空格
+			utcString = utcString.replace(" -", "-");
+			if(':' != utcString.charAt(utcString.length() - 3)){
+				utcString = utcString.substring(0, utcString.length() - 2) + ":00";
+			}
+
+			if (StrUtil.contains(utcString, CharUtil.DOT)) {
+				// 带毫秒，格式类似：2018-09-13T05:34:31.999-08:00
+				return new DateTime(utcString, DatePattern.UTC_MS_WITH_XXX_OFFSET_FORMAT);
+			} else {
+				// 格式类似：2018-09-13T05:34:31-08:00
+				return new DateTime(utcString, DatePattern.UTC_WITH_XXX_OFFSET_FORMAT);
+			}
 		} else {
 			if (length == DatePattern.UTC_SIMPLE_PATTERN.length() - 2) {
 				// 格式类似：2018-09-13T05:34:31